Split moq-json into snapshot (lossy) and stream (lossless) modules
Reshape moq-json / @moq/json around two explicit modes so the lossy vs
lossless distinction is unmissable:

- snapshot / Snapshot: lossy latest-value. One JSON value updated over
  time; a consumer only ever sees the most recent value (backlog
  collapses, older groups drop). This is the existing RFC 7396
  merge-patch snapshot/delta code, moved out of the crate/package root.
- stream / Stream: lossless append-log. Every record is preserved and
  delivered in order, nothing superseded.

Rust: the old crate root moves to a `snapshot` module; `lib.rs` keeps
the shared Error/Result/diff. TS: root Producer/Consumer move under a
`Snapshot` namespace, alongside the existing `Stream`.

Consumers updated: moq-mux catalog, moq-boy, js/hang catalog, js/watch,
js/publish, demo/web, the telemetry example, plus docs and the crate-map
guides.

This is a breaking rename of the published moq-json / @moq/json surface.
moq-json is not in the dev-only list in Branch Targeting (wire /
moq-net / hang / ffi / wrappers), so it targets main.
