+++ doc/concept/layer/hang.md
@@ -80,7 +80,7 @@ This is the minimum amount of information required to initialize a video decoder
 The base catalog carries only the media sections (`video` and `audio`).
 Applications add their own root sections (for example `scte35`) without modifying hang.
 
-The catalog is a JSON document published through the merge-patch helper (`@moq/json` / `moq-json`), and an extension is just an extra top-level key:
+The catalog is a JSON document published through the merge-patch snapshot helper (the `Snapshot` mode of `@moq/json` / `moq-json`), and an extension is just an extra top-level key:
 
 - **Reading**: the base schema is permissive, so unknown sections pass through validation untouched.
   A base consumer ignores them; an extension reads its own section and treats its absence as "not present".
@@ -100,8 +100,8 @@ A custom catalog section can carry its payload inline (for low-rate metadata), o
 
 The `@moq/publish` and `@moq/watch` components publish and subscribe to these tracks generically, with no per-application support. Each exposes a low-level track hook and re-exports `@moq/json` so the application encodes the payload itself:
 
-- **Publish**: `broadcast.publishTrack(name, serve)` runs `serve(track, effect)` per subscriber. For JSON, serve each track from a shared track-less `Json.Producer` (the same fan-out producer the catalog uses, seeding late joiners with the latest value). Advertise the track by writing your own catalog section with `broadcast.catalog.mutate(...)`.
-- **Watch**: `broadcast.subscribeTrack(name, priority, consume)` follows the active broadcast across reconnects. For JSON, wrap the track in a `Json.Consumer` inside `consume`. Read your section back from `broadcast.catalog` (unknown sections pass through the loose schema).
+- **Publish**: `broadcast.publishTrack(name, serve)` runs `serve(track, effect)` per subscriber. For JSON, serve each track from a shared track-less `Json.Snapshot.Producer` (the same fan-out producer the catalog uses, seeding late joiners with the latest value). Advertise the track by writing your own catalog section with `broadcast.catalog.mutate(...)`.
+- **Watch**: `broadcast.subscribeTrack(name, priority, consume)` follows the active broadcast across reconnects. For JSON, wrap the track in a `Json.Snapshot.Consumer` inside `consume`. Read your section back from `broadcast.catalog` (unknown sections pass through the loose schema).
 
 So an application supports something like SCTE-35 entirely in its own code: publish an `scte35` section (and optionally a track) on one side, read it on the other, without hang, `@moq/publish`, or `@moq/watch` knowing anything about SCTE-35.
