+++ doc/lib/js/@moq/publish.md
@@ -136,7 +136,7 @@ Beyond audio and video, you can publish arbitrary application tracks within the
 same broadcast (no separate broadcast needed). `publishTrack(name, serve)` runs
 `serve(track, effect)` for each subscriber; it rejects the built-in track names
 (catalog/audio/video). Encode the payload yourself with the re-exported
-`@moq/json`: a track-less `Json.Producer` is the same fan-out producer the catalog
+`@moq/json`: a track-less `Json.Snapshot.Producer` is the same fan-out producer the catalog
 uses, seeding late joiners with the latest value.
 
 `publishTrack` does not touch the catalog; advertise the track by writing your own
@@ -147,7 +147,7 @@ like an `scte35` section with no hang-specific support:
 ```typescript
 import { Json } from "@moq/publish";
 
-const scte35 = new Json.Producer<{ splices: number[] }>({ initial: { splices: [] } });
+const scte35 = new Json.Snapshot.Producer<{ splices: number[] }>({ initial: { splices: [] } });
 broadcast.publishTrack("scte35.json", (track, effect) => scte35.serve(track, effect));
 broadcast.catalog.mutate((c) => {
     c.scte35 = { track: "scte35.json" };
