+++ js/json/src/index.ts
@@ -1,12 +1,17 @@
 /**
- * Snapshot/delta JSON publishing over MoQ tracks using RFC 7396 JSON Merge Patch. A
- * {@link Producer} writes a JSON value to one track or fans it out to many; a {@link Consumer}
- * reconstructs the value on the other side.
+ * JSON publishing over MoQ tracks, in two modes:
+ *
+ * - {@link Snapshot}: **lossy**. One JSON value updated over time; a consumer only gets the most
+ *   recent value. Intermediate updates are collapsed and older groups are dropped.
+ * - {@link Stream}: **lossless**. An ordered append-log of self-contained records; every record
+ *   is preserved and delivered in order, nothing is ever superseded.
+ *
+ * Pick {@link Snapshot} when consumers care about "what is the value now" (a catalog, a status
+ * document) and {@link Stream} when they care about every record (an event log, a media timeline).
  *
  * @module
  */
 
-export { Consumer } from "./consumer.ts";
 export { type Diff, deepEqual, diff, merge } from "./diff.ts";
-export { type Config, Producer } from "./producer.ts";
+export * as Snapshot from "./snapshot/index.ts";
 export * as Stream from "./stream.ts";
