+++ js/json/src/snapshot/compression.test.ts
