+++ js/json/src/snapshot/producer.test.ts
