+++ rs/moq-mux/src/catalog/producer.rs
@@ -22,8 +22,8 @@ use super::hang::{Catalog, CatalogExt, Consumer, Extra};
 /// group (deltas disabled). This routes catalog publishing through the JSON merge-patch helper
 /// so deltas can be enabled later without changing the wire format used today.
 pub struct Producer<E: CatalogExt = ()> {
-	hang: moq_json::Producer<Catalog<E>>,
-	hangz: moq_json::Producer<Catalog<E>>,
+	hang: moq_json::snapshot::Producer<Catalog<E>>,
+	hangz: moq_json::snapshot::Producer<Catalog<E>>,
 	msf_track: moq_net::TrackProducer,
 
 	current: Arc<Mutex<Catalog<E>>>,
@@ -80,14 +80,14 @@ impl<E: CatalogExt> Producer<E> {
 		let msf_track = broadcast.create_track(moq_net::Track::new(moq_msf::DEFAULT_NAME))?;
 
 		// Disable deltas for now to stay byte-compatible with consumers that only read snapshots.
-		let mut json_config = moq_json::ProducerConfig::default();
+		let mut json_config = moq_json::snapshot::ProducerConfig::default();
 		json_config.delta_ratio = 0;
-		let hang = moq_json::Producer::new(hang_track, json_config.clone());
+		let hang = moq_json::snapshot::Producer::new(hang_track, json_config.clone());
 
 		// The `.z` track carries the same catalog, DEFLATE-compressed. Deltas stay off for parity
 		// with the plaintext track; only the per-group compression differs.
 		json_config.compression = true;
-		let hangz = moq_json::Producer::new(hangz_track, json_config);
+		let hangz = moq_json::snapshot::Producer::new(hangz_track, json_config);
 
 		Ok(Self {
 			hang,
@@ -209,8 +209,8 @@ impl<E: CatalogExt> Producer<E> {
 /// On drop, the hang, compressed-hang, and MSF catalog tracks are updated if the catalog was mutated.
 pub struct Guard<'a, E: CatalogExt = ()> {
 	catalog: MutexGuard<'a, Catalog<E>>,
-	hang: &'a mut moq_json::Producer<Catalog<E>>,
-	hangz: &'a mut moq_json::Producer<Catalog<E>>,
+	hang: &'a mut moq_json::snapshot::Producer<Catalog<E>>,
+	hangz: &'a mut moq_json::snapshot::Producer<Catalog<E>>,
 	msf_track: &'a mut moq_net::TrackProducer,
 	updated: bool,
 }
